refactor(v0.52 phase 2): cli/ui/ → ui/ top-level (cross-process 공유)
ui/는 이제 cli (event_renderer로 IPC events 렌더링)와 server (agentic_ui
emit_*로 IPC events 발신) 모두 import 가능한 cross-process capability.

이동 (8 파일 + __init__):
- core/cli/ui/{agentic_ui,console,event_renderer,mascot,panels,status,
  tool_tracker}.py → core/ui/<same>

Import 경로 갱신: 28 파일 (core 14 + tests 14)
- `from core.cli.ui.<X> import` → `from core.ui.<X> import`

이 분리의 의미:
- 향후 phase 4에서 server/가 신설되면 server-side handlers가 core.ui의
  emit_*를 import해서 thin client 렌더러에 events 전달
- cli/와 server/가 둘 다 ui/만 의존 (양방향 import 금지) → 깔끔한 import
  방향 (phase 6 import-linter로 강제)

검증:
- ruff + format clean (13 import block 정렬 자동수정)
- mypy strict clean (227 modules, ui top-level 승격으로 -1)
- pytest 4075 passed (1 skipped, 19 deselected)
- E2E `geode analyze "Cowboy Bebop" --dry-run` → A (68.4) unchanged

다음 phase: cli/__init__:_thin_interactive_loop → cli/repl.py 분리,
COMMAND_REGISTRY (RunLocation) 신설, **OAuth flow CLI 직접 실행 (B1, B3 정식 해소)**
