+++ core/cli/report_renderer.py
@@ -13,11 +13,11 @@ import logging
 from pathlib import Path
 from typing import Any
 
-from core.cli.ui.console import console
-from core.cli.ui.status import GeodeStatus
 from core.config import settings
 from core.paths import get_project_root
 from core.skills.reports import ReportFormat, ReportGenerator, ReportTemplate
+from core.ui.console import console
+from core.ui.status import GeodeStatus
 
 log = logging.getLogger(__name__)
