+++ core/cli/startup.py
@@ -17,9 +17,9 @@ from pathlib import Path
 
 from core.cli._helpers import mask_key as _mask_key
 from core.cli._helpers import upsert_env as _upsert_env
-from core.cli.ui.console import console
 from core.config import settings
 from core.memory.project import ProjectMemory
+from core.ui.console import console
 
 log = logging.getLogger(__name__)
