+++ core/graph.py
@@ -356,7 +356,7 @@ def _skip_check_node(state: GeodeState) -> dict[str, Any]:
     """
     skip_nodes = state.get("skip_nodes", [])
     if "verification" in skip_nodes:
-        from core.cli.ui.agentic_ui import emit_node_skipped
+        from core.ui.agentic_ui import emit_node_skipped
 
         emit_node_skipped("verification", "Dynamic Graph skip_nodes")
         log.info("Dynamic Graph: verification skipped (in skip_nodes)")
@@ -565,7 +565,7 @@ def build_graph(
                 iteration,
                 max_iter,
             )
-            from core.cli.ui.agentic_ui import emit_feedback_loop
+            from core.ui.agentic_ui import emit_feedback_loop
 
             emit_feedback_loop(iteration, conf_normalized * 100, confidence_threshold * 100)
             return "gather"
@@ -599,7 +599,7 @@ def build_graph(
             )
             return "synthesizer"
 
-        from core.cli.ui.agentic_ui import emit_feedback_loop
+        from core.ui.agentic_ui import emit_feedback_loop
 
         emit_feedback_loop(iteration, conf_normalized * 100, effective_threshold * 100)
         log.info(
